# references/TheBrain_API_Docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TheBrain API",
    "description": "Welcome to TheBrain API. You can try out the API right here on this page:\r\n\r\n* <a href=\"https://app.thebrain.com/api-keys\" target=\"_blank\">Generate an API Key</a>.\r\n* Return to this page and click the `Authorize` button below these instructions, paste your key, click `Authorize` then `Close`.\r\n* Click an endpoint to view its documentation.\r\n* Click the `Try it out` button, fill out the required items and click the 'Execute` button.\r\n* The resulting request and response are displayed.\r\n\r\nYou can also see example projects for <a href=\"https://github.com/TheBrainTech/thebrain-api-quickstart-blazor\" target=\"_blank\">Blazor</a>,\r\n<a href=\"https://github.com/TheBrainTech/thebrain-api-quickstart-python\" target=\"_blank\">Python</a>,\r\nand <a href=\"https://github.com/TheBrainTech/thebrain-api-quickstart-node\" target=\"_blank\">Node</a>\r\non GitHub.",
    "version": "1.0.5"
  },
  "paths": {
...
    "/links/{brainId}": {
      "post": {
        "tags": [
          "Links"
        ],
        "summary": "Creates a link",
        "description": "Link properties and expected values:\r\n            \r\n    Guid 'thoughtIdA' - The ID of the thought to begin the link from\r\n    Guid 'thoughtIdB' - The ID of the thought to terminate the link at\r\n            \r\n    string 'name' (Optional) - The label for the link\r\n            \r\n    int 'relation' - The relationship of the link, relative from 'thoughtIdA' to 'thoughtIdB'\r\n         Child = 1\r\n         Parent = 2\r\n         Jump = 3\r\n         Sibling = 4",
        "operationId": "createLink",
        "parameters": [
          {
            "name": "brainId",
            "in": "path",
            "description": "The ID of the brain",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Details of the link to create",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCreateModel"
              },
              "examples": {
                "Link": {
                  "value": {"thoughtIdA":"2c357ff8-2b29-4e3e-ab93-3502277df4fd","thoughtIdB":"cb2a43ad-2230-4a22-94fe-f8d932fe8d82","relation":1,"name":"New Link"}
                }
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCreateModel"
              },
              "examples": {
                "Link": {
                  "value": {"thoughtIdA":"2c357ff8-2b29-4e3e-ab93-3502277df4fd","thoughtIdB":"cb2a43ad-2230-4a22-94fe-f8d932fe8d82","relation":1,"name":"New Link"}
                }
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCreateModel"
              },
              "examples": {
                "Link": {
                  "value": {"thoughtIdA":"2c357ff8-2b29-4e3e-ab93-3502277df4fd","thoughtIdB":"cb2a43ad-2230-4a22-94fe-f8d932fe8d82","relation":1,"name":"New Link"}
                }
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCreateModel"
              },
              "examples": {
                "Link": {
                  "value": {"thoughtIdA":"2c357ff8-2b29-4e3e-ab93-3502277df4fd","thoughtIdB":"cb2a43ad-2230-4a22-94fe-f8d932fe8d82","relation":1,"name":"New Link"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success - returns the newly created link ID",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLinkResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLinkResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLinkResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "If the link can't be created"
          },
          "401": {
            "description": "If the API Key is missing or invalid"
          },
          "404": {
            "description": "If the source or target thought isn't found"
          }
        }
      }
    },
...
    "/thoughts/{brainId}": {
      "get": {
        "tags": [
          "Thoughts"
        ],
        "summary": "Returns the first thought matching the name exactly",
        "operationId": "getThoughtByName",
        "parameters": [
          {
            "name": "brainId",
            "in": "path",
            "description": "The ID of the brain",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "nameExact",
            "in": "query",
            "description": "The exact name of the thought",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the first thought matching the name exactly.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ThoughtDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThoughtDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThoughtDto"
                }
              }
            }
          },
          "401": {
            "description": "If the API Key is missing or invalid"
          }
        }
      },
      "post": {
        "tags": [
          "Thoughts"
        ],
        "summary": "Creates a thought",
        "description": "Thought properties and expected values:\r\n            \r\n    string 'name' - The Name of the thought\r\n    string 'label' (Optional) - The Label of the thought\r\n\r\n    Guid 'sourceThoughtId' (Optional)  - The ID of source thought. Exclude to create an Orphan thought\r\n    Guid 'typeId' (Optional)  - The ID of the thought type\r\n            \r\n    int 'acType' - The access type of the thought\r\n         Public = 0\r\n         Private = 1\r\n            \r\n    int 'kind' - The kind of thought\r\n         Normal = 1\r\n         Type = 2\r\n         Event = 3\r\n         Tag = 4\r\n         System = 5\r\n            \r\n    int 'relation' - The relationship of the created thought to the source thought\r\n         Child = 1\r\n         Parent = 2\r\n         Jump = 3\r\n         Sibling = 4",
        "operationId": "createThought",
        "parameters": [
          {
            "name": "brainId",
            "in": "path",
            "description": "The ID of the brain",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Details of the thought to create",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ThoughtCreateModel"
              },
              "examples": {
                "Thought": {
                  "value": {"name":"New Thought","kind":1,"label":"New Label","sourceThoughtId":"108883a9-86a4-4c88-9017-891918860153","relation":1,"acType":0}
                }
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThoughtCreateModel"
              },
              "examples": {
                "Thought": {
                  "value": {"name":"New Thought","kind":1,"label":"New Label","sourceThoughtId":"108883a9-86a4-4c88-9017-891918860153","relation":1,"acType":0}
                }
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ThoughtCreateModel"
              },
              "examples": {
                "Thought": {
                  "value": {"name":"New Thought","kind":1,"label":"New Label","sourceThoughtId":"108883a9-86a4-4c88-9017-891918860153","relation":1,"acType":0}
                }
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ThoughtCreateModel"
              },
              "examples": {
                "Thought": {
                  "value": {"name":"New Thought","kind":1,"label":"New Label","sourceThoughtId":"108883a9-86a4-4c88-9017-891918860153","relation":1,"acType":0}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success - returns the ID of the newly created thought.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateThoughtResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateThoughtResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateThoughtResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "If the creation of the thought failed"
          },
          "401": {
            "description": "If the API Key is missing or invalid"
          },
          "404": {
            "description": "If the 'sourceId' thought was not found"
          }
        }
      }
    },
...
    "/thoughts/{brainId}/pins": {
      "get": {
        "tags": [
          "Thoughts"
        ],
        "summary": "Gets a list of all pinned thoughts in a brain",
        "operationId": "getPinnedThoughts",
        "parameters": [
          {
            "name": "brainId",
            "in": "path",
            "description": "The ID of the brain",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If the list of pinned thoughts failed to be retrieved"
          },
          "401": {
            "description": "If the API Key is missing or invalid"
          }
        }
      }
    },
...
    "/thoughts/{brainId}/types": {
      "get": {
        "tags": [
          "Thoughts"
        ],
        "summary": "Gets a list of all Types in a specified brain",
        "description": "Thought properties and values:\r\n            \r\n    Guid 'id' - The ID of the thought\r\n    Guid 'brainId' - The ID of the brain containing the thought\r\n    Guid 'typeId' - The ID of the thought type\r\n            \r\n    DateTime 'creationDateTime' - The DateTime that the thought was created at\r\n    DateTime 'modificationDateTime' - The DateTime that the thought was last modified at\r\n    DateTime 'forgottenDateTime' - The DateTime that the thought was forgotten at\r\n    DateTime 'linksModificationDateTime' - The DateTime that the thought's links were last modified at\r\n\r\n    string 'name' - The Name of the thought\r\n    string 'label' - The Label of the thought\r\n    string 'foregroundColor' - The RGB hexadecimal color of the thought's foreground color (Ex: '#ff7145')\r\n    string 'backgroundColor' - The RGB hexadecimal color of the thought's background color (Ex: '#ff7145')\r\n            \r\n    int 'acType' - The access type of the thought\r\n         Public = 0\r\n         Private = 1\r\n            \r\n    int 'kind' - The kind of thought\r\n         Normal = 1\r\n         Type = 2\r\n         Event = 3\r\n         Tag = 4\r\n         System = 5",
        "operationId": "getTypes",
        "parameters": [
          {
            "name": "brainId",
            "in": "path",
            "description": "The ID of the brain",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If the list of Types failed to be retrieved"
          },
          "401": {
            "description": "If the API Key is missing or invalid"
          }
        }
      }
    },
    "/thoughts/{brainId}/tags": {
      "get": {
        "tags": [
          "Thoughts"
        ],
        "summary": "Gets a list of all Tags in a specified brain",
        "description": "Thought properties and values:\r\n            \r\n    Guid 'id' - The ID of the thought\r\n    Guid 'brainId' - The ID of the brain containing the thought\r\n    Guid 'typeId' - The ID of the thought type\r\n            \r\n    DateTime 'creationDateTime' - The DateTime that the thought was created at\r\n    DateTime 'modificationDateTime' - The DateTime that the thought was last modified at\r\n    DateTime 'forgottenDateTime' - The DateTime that the thought was forgotten at\r\n    DateTime 'linksModificationDateTime' - The DateTime that the thought's links were last modified at\r\n\r\n    string 'name' - The Name of the thought\r\n    string 'label' - The Label of the thought\r\n    string 'foregroundColor' - The RGB hexadecimal color of the thought's foreground color (Ex: '#ff7145')\r\n    string 'backgroundColor' - The RGB hexadecimal color of the thought's background color (Ex: '#ff7145')\r\n            \r\n    int 'acType' - The access type of the thought\r\n         Public = 0\r\n         Private = 1\r\n            \r\n    int 'kind' - The kind of thought\r\n         Normal = 1\r\n         Type = 2\r\n         Event = 3\r\n         Tag = 4\r\n         System = 5",
        "operationId": "getTags",
        "parameters": [
          {
            "name": "brainId",
            "in": "path",
            "description": "The ID of the brain",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThoughtDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If the list of Tags failed to be retrieved"
          },
          "401": {
            "description": "If the API Key is missing or invalid"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLinkResponseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateThoughtResponseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LinkCreateModel": {
        "type": "object",
        "properties": {
          "thoughtIdA": {
            "type": "string",
            "format": "uuid"
          },
          "thoughtIdB": {
            "type": "string",
            "format": "uuid"
          },
          "relation": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ThoughtCreateModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "kind": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "sourceThoughtId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "relation": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "acType": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ThoughtDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "brainId": {
            "type": "string",
            "format": "uuid"
          },
          "creationDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "modificationDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cleanedUpName": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "displayModificationDateTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "forgottenDateTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "linksModificationDateTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "acType": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "type": "integer",
            "format": "int32"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "foregroundColor": {
            "type": "string",
            "nullable": true
          },
          "backgroundColor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {apiKey}\"",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}